task_parser/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
python-dotenv
sqlalchemy
asyncpg
psycopg2-binary